ipmi: run sol deactivate on shutdown.
This makes it more likely that the controller will be available for
another connection quickly. Repeated connections will still eventually
cause the console to stop accepting connections for a short while, but
on some obms (e.g. our old sun blades), not doing this causes an
immediate failure on the next connection.